requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-admin==2.0.1
django-bootstrap4==2.3.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-rest-framework==0.1.0
django-six==1.0.4
django-tastypie==0.14.3
djangorestframework==3.12.2
idna==2.10
MLB-StatsAPI==1.0.1
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2021.1
requests==2.25.1
screen==1.0.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
xlwt==1.3.0
